global gickup UserAgent
